requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
stringzilla==4.4.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.22
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
